pyproject.toml
[tool.poetry]
name = "langchain-rag-project"
version = "0.1.0"
description = "A project using LangChain, RAG, and Llama model for document querying"
authors = ["Nancy Zhao <szhao@hr.lacounty.gov>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "1.0.1"
langchain = "0.2.2"
langchain-community = "0.2.3"
langchain-openai = "0.1.8"
chromadb = "0.5.0"
openai = "1.31.1"
tiktoken = "0.7.0"
llama-cpp-python = "^0.1.48"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"
pandas = "^1.5.3"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"